XNAT-6990 Custom routing rules by receiver

+++ src/main/java/org/nrg/dcm/scp/exceptions/DICOMReceiverWithDuplicateIdException.java
@@ -32,7 +32,7 @@ public class DICOMReceiverWithDuplicateIdException extends DICOMReceiverWithDupl
 
     @SuppressWarnings("unused")
     public DICOMReceiverWithDuplicateIdException(@Nonnull final DicomSCPInstance instance) {
-        this(instance.getId());
+        this( (int) instance.getId());
     }
 
     @SuppressWarnings("unused")
